Move JAXB adapter to the impl package
